fix(tornado): linting
