+++ epsagon/utils.py
@@ -3,38 +3,8 @@ Utilities for Epsagon module.
 """
 
 from __future__ import absolute_import
-from six.moves import urllib
-
 from epsagon.constants import TRACE_COLLECTOR_URL, REGION
 from .trace import tracer
-
-# Method to URL dict.
-BLACKLIST_URLS = {
-    str.endswith: [
-        'epsagon.com',
-    ],
-    str.__contains__: [
-        'accounts.google.com',
-    ],
-}
-
-# Ignored content types for web frameworks.
-IGNORED_CONTENT_TYPES = [
-    'image',
-    'audio',
-    'video',
-    'font',
-    'zip',
-    'css',
-]
-IGNORED_FILE_TYPES = [
-    '.js',
-    '.jsx',
-    '.woff',
-    '.woff2',
-    '.ttf',
-    '.eot',
-]
 
 
 def add_data_if_needed(dictionary, name, data):
@@ -48,37 +18,6 @@ def add_data_if_needed(dictionary, name, data):
     dictionary[name] = None
     if not tracer.metadata_only:
         dictionary[name] = data
-
-
-def is_blacklisted_url(url):
-    """
-    Return whether the URL blacklisted or not.
-    Using BLACKLIST_URLS methods against the URLs.
-    :param url: url string
-    :return: True if URL is blacklisted, else False
-    """
-
-    url = urllib.parse.urlparse(url).netloc
-
-    for method in BLACKLIST_URLS:
-        for blacklist_url in BLACKLIST_URLS[method]:
-            if method(url, blacklist_url):
-                return True
-    return False
-
-
-def ignore_request(content, path):
-    """
-    Return true if HTTP request in web frameworks should be omitted.
-    :param content: accept mimetype header
-    :param path: request path
-    :return: Bool
-    """
-
-    return (
-        any([x in content for x in IGNORED_CONTENT_TYPES]) or
-        any([path.endswith(x) for x in IGNORED_FILE_TYPES])
-    )
 
 
 def get_tc_url(use_ssl):
