+++ tests/test_utils.py
@@ -1,5 +1,6 @@
 import epsagon.trace
 import epsagon.utils
+import epsagon.wrappers.http_filters
 from epsagon.trace import tracer
 
 
@@ -13,7 +14,7 @@ def test_blacklist_url():
     :return: None
     """
 
-    epsagon.utils.BLACKLIST_URLS = {
+    epsagon.wrappers.http_filters.BLACKLIST_URLS = {
         str.endswith: [
             '.com',
         ],
@@ -22,11 +23,11 @@ def test_blacklist_url():
         ],
     }
 
-    assert epsagon.utils.is_blacklisted_url('http://www.google.com')
-    assert epsagon.utils.is_blacklisted_url('https://www.restricted-site.org')
-    assert epsagon.utils.is_blacklisted_url('http://www.restricted-site.com')
-    assert not epsagon.utils.is_blacklisted_url('https://www.com.org')
-    assert not epsagon.utils.is_blacklisted_url('http://www.google.org')
+    assert epsagon.wrappers.http_filters.is_blacklisted_url('http://www.google.com')
+    assert epsagon.wrappers.http_filters.is_blacklisted_url('https://www.restricted-site.org')
+    assert epsagon.wrappers.http_filters.is_blacklisted_url('http://www.restricted-site.com')
+    assert not epsagon.wrappers.http_filters.is_blacklisted_url('https://www.com.org')
+    assert not epsagon.wrappers.http_filters.is_blacklisted_url('http://www.google.org')
 
 
 def test_original_blacklist_url():
@@ -35,5 +36,5 @@ def test_original_blacklist_url():
     :return: None
     """
 
-    assert epsagon.utils.is_blacklisted_url('http://tc.us-east-1.epsagon.com')
-    assert epsagon.utils.is_blacklisted_url('https://client.tc.epsagon.com')
+    assert epsagon.wrappers.http_filters.is_blacklisted_url('http://tc.us-east-1.epsagon.com')
+    assert epsagon.wrappers.http_filters.is_blacklisted_url('https://client.tc.epsagon.com')
